21 temmuz

+++ src/main/java/ticket/dataaccess/abstracts/TicketRepository.java
@@ -1,17 +1,18 @@
 package ticket.dataaccess.abstracts;
 
-import jakarta.persistence.metamodel.SingularAttribute;
+
 import org.springframework.data.jpa.domain.AbstractPersistable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ticket.core.utilities.results.DataResult;
 import ticket.entities.concretes.Ticket;
 
+import javax.persistence.metamodel.SingularAttribute;
 import java.io.Serializable;
 
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket,Integer > {
+public interface TicketRepository extends JpaRepository<Ticket,Long > {
 
     DataResult<Ticket>getTicketById(long id);
